change the names of fields and classes
