+++ src/main/java/MissingFieldException.java
@@ -1,4 +1,4 @@
-public class MissingFieldException extends DukeException{
+public class MissingFieldException extends IrisException {
     public MissingFieldException(String field) {
         super(field + " cannot be empty.");
     }
