+++ src/main/java/NoTaskException.java
@@ -1,4 +1,4 @@
-public class NoTaskException extends DukeException{
+public class NoTaskException extends IrisException {
     public NoTaskException() {
         super("I'm sorry, I don't know what you mean.");
     }
